refactor with exception
