+++ src/main/java/com/payper/server/auth/exception/AuthException.java
@@ -1,0 +1,15 @@
+package com.payper.server.auth.exception;
+
+import com.payper.server.global.response.ErrorCode;
+import lombok.Getter;
+import org.springframework.security.core.AuthenticationException;
+
+@Getter
+public class AuthException extends AuthenticationException {
+    private final ErrorCode errorCode;
+
+    public AuthException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
