+++ src/main/java/com/payper/server/auth/jwt/util/JwtRefreshTokenUtil.java
@@ -2,8 +2,6 @@ package com.payper.server.auth.jwt.util;
 
 import com.payper.server.auth.jwt.RefreshTokenRepository;
 import com.payper.server.auth.jwt.entity.RefreshTokenEntity;
-import com.payper.server.auth.jwt.exception.ReissueException;
-import com.payper.server.global.response.ErrorCode;
 import io.jsonwebtoken.Jwts;
 import jakarta.annotation.PostConstruct;
 import jakarta.servlet.http.Cookie;
